Add Ruby hello-world code
This is the sample code which performs a "ping" on Redis and prints
"Hello World!" if the method call did not give an error.

If this code is run on the command line inside the official Ruby
Docker image, it will launch WEBrick web server listening on the
localhost interface, port 4567.