Add OpenShift template (binary builds)
This OpenShift template creates all the required resource objects on the
container platform.

Build configurations in this revision expect binary builds. This is useful
to test everything is working without having to upload all files to a GitHub
repository. You'll need to run "oc start-build ..." to trigger builds and
deployments.