Add Gemfile and Rackup config
When using OpenShift Source-to-Image (S2I) build strategy, the easiest way to
run a Ruby web application is using Rack. We could override S2I scripts which
assemble and configure the runtime environment if we needed though.

So, we add a config.ru file for the rackup command and a Gemfile which
will install the required Ruby gems (redis, sinatra). And we adapt the
hello-world.rb to be compatible with config.ru configuration file.

Because the rackup command is set to listen on port 8080 we also update the
OpenShift template to expose 8080 on the container and set this on the service
as well.